+++ src/app/api/comments/[id]/route.ts
@@ -1,9 +1,10 @@
 import { NextResponse } from "next/server";
 import { getServerSession } from "next-auth";
-import { authOptions } from "@/app/api/auth/[...nextauth]/route";
+import { authOptions } from "@/app/utils/authOptions";
 import { prisma } from "@/lib/prisma";
 
-export async function DELETE(req: Request, { params }: { params: { id: string } }) {
+export async function DELETE(req: Request, { params }: { params: Promise<{ id: string }> }) {
+  const {id} = await params;
   const session = await getServerSession(authOptions);
 
   if (!session) {
@@ -11,7 +12,7 @@ export async function DELETE(req: Request, { params }: { params: { id: string } 
   }
 
   try {
-    await prisma.comment.delete({ where: { id: params.id } });
+    await prisma.comment.delete({ where: { id: id } });
 
     return NextResponse.json({ message: "Comment deleted successfully" });
   } catch (error) {
