+++ src/app/api/portfolio/edit/route.tsx
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from "next/server";
 import { getServerSession } from "next-auth";
-import { authOptions } from "../../auth/[...nextauth]/route";
+import { authOptions } from "@/app/utils/authOptions";
 import { PrismaClient } from "@prisma/client";
 import fs from "fs";
 import path from "path";
