+++ src/app/portfolio/edit/page.tsx
@@ -1,5 +1,5 @@
 import { getServerSession } from "next-auth";
-import { authOptions } from "../../api/auth/[...nextauth]/route";
+import { authOptions } from "@/app/utils/authOptions";
 import { redirect } from "next/navigation";
 import EditPortfolioForm from "./EditPortfolioForm";
