+++ src/app/utils/authOptions.ts
@@ -1,0 +1,43 @@
+import NextAuth, { AuthOptions } from "next-auth";
+import CredentialsProvider from "next-auth/providers/credentials";
+import { PrismaClient } from "@prisma/client";
+import { compare } from "bcryptjs";
+
+const prisma = new PrismaClient();
+
+export const authOptions: AuthOptions = {
+  providers: [
+    CredentialsProvider({
+      name: "Credentials",
+      credentials: {
+        username: { label: "Username", type: "text", placeholder: "your-username" },
+        password: { label: "Password", type: "password" },
+      },
+      async authorize(credentials) {
+        if (!credentials?.username || !credentials?.password) {
+          throw new Error("Username and password are required");
+        }
+
+        const user = await prisma.owner.findUnique({
+          where: { username: credentials.username },
+        });
+
+        if (!user) {
+          throw new Error("User not found");
+        }
+
+        const isValidPassword = await compare(credentials.password, user.password);
+
+        if (!isValidPassword) {
+          throw new Error("Invalid password");
+        }
+
+        return { id: String(user.id), name: user.username };
+      },
+    }),
+  ],
+  session: {
+    strategy: "jwt",
+  },
+  secret: process.env.NEXTAUTH_SECRET, // Set this in your .env file
+};
